make sure all scanner are working and add automatic importing from a dedicated folder
